requirements.txt
fastapi
pymysql
SQLAlchemy
uvicorn
python-dotenv
pydantic[email]
pydantic-settings
passlib[bcrypt]
pyJWT
cryptography
python-multipart